fcode message -This is the message to send to motion
Sends a NML message of the feedrate from the F word.
The feedrate is actually baked into motion commands.
But to get status of f Word at the right time, I'm adding
a specfic 'code status' message with the fcode value.

In linxcnc K is used for thread feed rate (pitch)
cover this with a feed message