qtvcp Fix possible runaway jog on increments change
if you were continuous jogging with screen buttons and change the incrments
with key shortcut (i) then the jog would continue forever.
Though no current sample adjusts the increments with a shortcut,
it certainly could be added and the problem would show.
Now we cancel any jogging on increment change.